fix: Changed operator so that it does a server side apply rather than create for resources. (#470)

+++ crates/k8s-operator/src/services/keycloak.rs
@@ -5,7 +5,7 @@ use crate::error::Error;
 use crate::operator::crd::BionicSpec;
 use k8s_openapi::api::apps::v1::Deployment;
 use k8s_openapi::api::core::v1::{ConfigMap, Secret, Service};
-use kube::api::{DeleteParams, ObjectMeta, PostParams};
+use kube::api::{DeleteParams, ObjectMeta, Patch, PatchParams, PostParams};
 use kube::{Api, Client};
 use serde_json::json;
 
@@ -15,7 +15,7 @@ pub const KEYCLOAK_NAME: &str = "keycloak";
 // We are using envoy to add security headers to all responses from the main application.
 pub async fn deploy(client: Client, spec: BionicSpec, namespace: &str) -> Result<(), Error> {
     // Put the envoy.yaml into a ConfigMap
-    let config_map = serde_json::from_value(serde_json::json!({
+    let config_map = serde_json::json!({
         "apiVersion": "v1",
         "kind": "ConfigMap",
         "metadata": {
@@ -25,10 +25,15 @@ pub async fn deploy(client: Client, spec: BionicSpec, namespace: &str) -> Result
         "data": {
             "realm.json": CONFIG_JSON,
         }
-    }))?;
+    });
 
     let api: Api<ConfigMap> = Api::namespaced(client.clone(), namespace);
-    api.create(&PostParams::default(), &config_map).await?;
+    api.patch(
+        KEYCLOAK_NAME,
+        &PatchParams::apply(crate::MANAGER),
+        &Patch::Apply(config_map),
+    )
+    .await?;
 
     // Keycloak
     deployment::deployment(
@@ -117,44 +122,58 @@ pub async fn deploy(client: Client, spec: BionicSpec, namespace: &str) -> Result
     )
     .await?;
 
-    let mut secret_data = BTreeMap::new();
-    secret_data.insert("admin-password".to_string(), super::database::rand_hex());
+    keycloak_secrets(namespace, client).await?;
 
-    let keycloak_secret = Secret {
-        metadata: ObjectMeta {
-            name: Some("keycloak-secrets".to_string()),
-            namespace: Some(namespace.to_string()),
-            ..ObjectMeta::default()
-        },
-        string_data: Some(secret_data),
-        ..Default::default()
-    };
+    Ok(())
+}
 
+async fn keycloak_secrets(namespace: &str, client: Client) -> Result<(), Error> {
     let secret_api: Api<Secret> = Api::namespaced(client, namespace);
-    secret_api
-        .create(&PostParams::default(), &keycloak_secret)
-        .await?;
-
+    let secret = secret_api.get("keycloak-secrets").await;
+    if secret.is_err() {
+        let mut secret_data = BTreeMap::new();
+        secret_data.insert("admin-password".to_string(), super::database::rand_hex());
+        let keycloak_secret = Secret {
+            metadata: ObjectMeta {
+                name: Some("keycloak-secrets".to_string()),
+                namespace: Some(namespace.to_string()),
+                ..ObjectMeta::default()
+            },
+            string_data: Some(secret_data),
+            ..Default::default()
+        };
+        secret_api
+            .create(&PostParams::default(), &keycloak_secret)
+            .await?;
+    }
     Ok(())
 }
 
 pub async fn delete(client: Client, namespace: &str) -> Result<(), Error> {
     // Remove deployments
     let api: Api<Deployment> = Api::namespaced(client.clone(), namespace);
-    api.delete(KEYCLOAK_NAME, &DeleteParams::default()).await?;
+    if api.get(KEYCLOAK_NAME).await.is_ok() {
+        api.delete(KEYCLOAK_NAME, &DeleteParams::default()).await?;
+    }
 
     // Remove services
     let api: Api<Service> = Api::namespaced(client.clone(), namespace);
-    api.delete(KEYCLOAK_NAME, &DeleteParams::default()).await?;
+    if api.get(KEYCLOAK_NAME).await.is_ok() {
+        api.delete(KEYCLOAK_NAME, &DeleteParams::default()).await?;
+    }
 
     // Remove configmaps
     let api: Api<ConfigMap> = Api::namespaced(client.clone(), namespace);
-    api.delete(KEYCLOAK_NAME, &DeleteParams::default()).await?;
+    if api.get(KEYCLOAK_NAME).await.is_ok() {
+        api.delete(KEYCLOAK_NAME, &DeleteParams::default()).await?;
+    }
 
     // Remove secrets
     let api: Api<Secret> = Api::namespaced(client, namespace);
-    api.delete("keycloak-secrets", &DeleteParams::default())
-        .await?;
+    if api.get("keycloak-secrets").await.is_ok() {
+        api.delete("keycloak-secrets", &DeleteParams::default())
+            .await?;
+    }
 
     Ok(())
 }
