+++ crates/k8s-operator/src/services/keycloak_db.rs
@@ -11,7 +11,14 @@ use kube::{
     Client,
 };
 
-pub async fn deploy(client: Client, namespace: &str) -> Result<String, Error> {
+pub async fn deploy(client: Client, namespace: &str) -> Result<Option<String>, Error> {
+    // If the cluster is already created then leave it alone.
+    let cluster_api: Api<Cluster> = Api::namespaced(client.clone(), namespace);
+    let cluster = cluster_api.get("keycloak-db-cluster").await;
+    if cluster.is_ok() {
+        return Ok(None);
+    }
+
     let database_password: String = rand_hex();
 
     let cluster = Cluster {
@@ -38,8 +45,6 @@ pub async fn deploy(client: Client, namespace: &str) -> Result<String, Error> {
             },
         },
     };
-
-    let cluster_api: Api<Cluster> = Api::namespaced(client.clone(), namespace);
     cluster_api.create(&PostParams::default(), &cluster).await?;
 
     let mut secret_data = BTreeMap::new();
@@ -61,19 +66,23 @@ pub async fn deploy(client: Client, namespace: &str) -> Result<String, Error> {
         .create(&PostParams::default(), &dbowner_secret)
         .await?;
 
-    Ok(database_password)
+    Ok(Some(database_password))
 }
 
 pub async fn delete(client: Client, namespace: &str) -> Result<(), Error> {
     // Remove deployments
     let api: Api<Cluster> = Api::namespaced(client.clone(), namespace);
-    api.delete("keycloak-db-cluster", &DeleteParams::default())
-        .await?;
+    if api.get("keycloak-db-cluster").await.is_ok() {
+        api.delete("keycloak-db-cluster", &DeleteParams::default())
+            .await?;
+    }
 
     let secret_api: Api<Secret> = Api::namespaced(client, namespace);
-    secret_api
-        .delete("keycloak-db-owner", &DeleteParams::default())
-        .await?;
+    if api.get("keycloak-db-owner").await.is_ok() {
+        secret_api
+            .delete("keycloak-db-owner", &DeleteParams::default())
+            .await?;
+    }
 
     Ok(())
 }
